refactor([issue-756]): split CalendarTab into calendar/ submodules
The 1269-LOC CalendarTab.jsx is broken into a calendar/ subdirectory:
pure date/grid math (lifeGridMath.js, now unit-tested), the grid renderer
(LifeGrid.jsx), TimeStats, AddActivityForm, LifeEventsPanel, the icon map,
and the persisted-prefs hook. CalendarTab.jsx is now a 219-LOC orchestrator.

Also fixes a latent reference bug in handleRemoveEvent: it read an
out-of-scope lifeEvents variable; it now resolves the removed event's name
from data.events.
